Final Commit
Ready for review. Renamed the files to their proper names. Added in the javascript file to the html file. Put the javascript file in strict mode. Stored a date in a variable. Stored the hour and month of that date into a variable. Stored an equation in a variable. Calculated the right image map to display. Added in the image to the html file. Set the date at the top to the correct date.